+++ output/Material_t1.csv
@@ -1,101 +1,10 @@
 id,material_type,material
-ac6137ca-0696-4177-9bef-21e9a5f4ef32,metal,Copper
-8e42d702-7187-4d3e-b081-37c1451323ae,metal,Steel
-4e93b273-1359-4f7c-aaa4-c58718dffbe4,plastic,PVC
-88b36bc8-dc0e-4fdb-bcc7-f1dca8267bbc,wood,Teak
-4464c451-3bd4-42c8-956f-9a39d2ddac30,plastic,PE
-e7f9dbed-d001-43fc-b632-138faadcf6a1,wood,Teak
-46428d95-1d45-4286-b88d-9b6bd13f5bda,wood,Teak
-587256f9-135d-40a9-af32-b5a74d783b74,metal,Copper
-2d98ecef-ec75-4cd8-b00b-dc461fd9d947,metal,Steel
-f832334a-e14d-4227-aa84-2fb3073ced1d,plastic,PE
-3acef501-7c2e-4684-b187-255ad98285a1,metal,Aluminium
-2eb11e1c-6adc-4b93-b963-d0f62f611594,metal,Aluminium
-a18e79f1-0d1b-40e4-a28c-0c36d96ae74b,plastic,PC
-b65d4593-6561-4997-889b-cac915262959,plastic,PC
-86406e3a-eb3c-40e9-8b60-091dffbcd407,plastic,PE
-38db8ee7-d723-43ac-99f5-1bcd1650e0c8,metal,Steel
-328ba502-ff50-4dc2-bebc-99f987a9ee9a,plastic,PVC
-b5e7d4d7-ab91-41ba-82fd-51c9e0f3d419,wood,Oak
-e07dd1a4-7330-4625-a176-47a502d4ae07,wood,Teak
-4835d67c-c327-4f4e-a77e-525b225472fa,plastic,PC
-213c2c89-583b-4c6b-8ce4-f84ca8dcf69c,wood,Teak
-fa09aa3e-ed4a-4b58-9e58-a02f19f81d23,metal,Copper
-17dd5d3d-0297-49ea-afc7-e39b01bac44d,plastic,PE
-139b5648-ba2e-4625-9ab6-b3305ecfe466,plastic,PVC
-8c7f12f5-1382-405c-9362-e97bd6035d5c,metal,Steel
-7561e5de-8e9b-4aab-a2a3-500f0d8e7aa2,metal,Aluminium
-7107d4d9-ce22-4b03-be7e-a216d9301d0c,wood,Oak
-b52adb78-40e8-4719-a0ff-ea3a2b2a3d4d,wood,Oak
-cee9db8d-0cf0-416d-8148-28d5fdbc2f6b,wood,Pine
-ea28bb21-e5d6-4430-b755-4aaa248cbc54,plastic,PE
-f3eb9e9d-f575-42a8-9db7-4df8d168af12,wood,Teak
-fa9ee2ef-2d13-45cd-82b5-79dee1d15700,wood,Oak
-537dc44f-da4c-4850-8527-57d17f3b695a,plastic,PE
-37ab7669-067f-47f8-914e-65c0bb1a2d74,metal,Aluminium
-54399311-06d0-4f13-bb4e-c5048cec09c3,plastic,PVC
-8e8a56b4-0192-4b72-990f-767743eaef7a,plastic,PE
-ce074695-1fe1-494b-8f15-3bffcf299c68,wood,Oak
-41f22a80-3f3f-4358-8e73-65ff271ad3bd,metal,Steel
-36ffb705-6b46-4ca4-9900-6506d1d4836f,plastic,PE
-7b80546e-77b3-4153-88ef-ebf394481cc7,wood,Oak
-f0e3404a-41c2-4e5f-bc0f-b8e250c72e98,metal,Steel
-72ce68a7-9ffb-455c-8e4f-e7f575d03113,plastic,PC
-db2a3526-ca76-4a11-9272-a713e7e59506,plastic,PVC
-0db29a8b-9d12-4e9b-a2a4-8363ebf0d252,metal,Steel
-d271e2cb-9cf2-42a9-980d-016d57ee31f4,wood,Pine
-5d1a2460-9a1c-47ac-9a6b-d23ab6a71152,metal,Copper
-be4a47ab-5d08-49c0-a13a-ae6e40357017,wood,Oak
-e3e43ef7-a2e9-47f2-96dc-14b73b47c493,plastic,PC
-e1cf0890-72a9-447d-bdf5-ae5d36ea673d,metal,Steel
-971bdd98-9d69-494c-ade0-9d18eccb7717,plastic,PE
-edd3bdcc-7146-49be-858e-65bfd8493bcf,metal,Aluminium
-97a78367-5fd9-4acd-bbae-a73f97695583,plastic,PE
-92bc454d-a6f4-445e-87b9-6313607d7b27,metal,Copper
-a78ea92c-5af0-4eac-a561-ab109b8fb32e,metal,Aluminium
-7ed9a0c4-338f-4f4a-bc66-acb37a08df13,metal,Steel
-bcc8b369-2ac2-4357-a801-6715fcc25d0f,wood,Pine
-98eb51bb-8002-4243-9811-db2aaff58d2f,metal,Aluminium
-3097f673-b247-4f99-ba92-b066e4f7e949,wood,Pine
-e007fbc4-1e93-46d9-8b81-af5111069fdc,plastic,PVC
-87e09a5c-fba4-4831-b7fe-4828d9c51e87,metal,Aluminium
-a1aef1e3-ed6a-458d-9238-45871c9b586f,metal,Steel
-2151d208-8693-4fd9-805f-52736a41fba3,wood,Oak
-f893bf9f-fc37-48a9-84d9-96a2b1308a8c,metal,Aluminium
-7320b6c3-b74c-4a9e-bc9f-aa76c087141c,plastic,PC
-e391a3a6-d603-47af-aced-a0e653f9c9c8,wood,Teak
-286dec7c-8ef0-4186-8456-3420986f2412,plastic,PVC
-3862aca4-086f-43c0-8fde-c82ceb59df52,wood,Pine
-f1dfb248-00be-4e76-81e1-6204f665a803,plastic,PC
-1cf25398-39ca-4109-ab22-e37a46cbf246,wood,Pine
-1781be2b-13aa-4ab9-b801-8a3833b46a1b,plastic,PC
-82e8eea1-82d4-4abc-a443-21529a2787b8,plastic,PE
-c45403e6-2285-4382-8477-f07473fdd156,metal,Aluminium
-893e691f-2431-42b8-831b-7f44f2a9c242,plastic,PC
-f51f13c8-b475-42cb-8b76-8e43cc3f7bd2,metal,Copper
-d95f5ae2-2bb2-461d-926a-efdbcb3a6782,metal,Steel
-8ec3b119-10a3-409e-a370-a50a6ca3f521,metal,Aluminium
-8bf5e1b7-967e-4f9d-9076-9436bac798cc,plastic,PVC
-26c47b89-76f7-4788-96ee-b1fc9930063f,metal,Steel
-980b0d3b-cdac-4b8d-b212-f81df0e97e92,plastic,PC
-c60b3523-bb43-4cad-97ad-345917e0059a,plastic,PVC
-cf83d1ec-ba7f-4593-958a-df3edae8c834,metal,Copper
-1bad0e0f-8ac0-460b-9142-0d4e7ca4ae1c,wood,Oak
-f3c33f5e-71dd-45f2-bc36-8151d3f625bf,wood,Oak
-337147cc-188f-4dda-befe-292b42e88d70,wood,Teak
-3945d713-74e9-4f15-b960-4db7453e3391,plastic,PVC
-dccf7856-94fe-4e08-8714-148b6f337305,wood,Teak
-e576011c-9820-4a61-b9a1-e6b013f4bb73,plastic,PVC
-45337d09-34ab-4b4c-9a3e-f2c18ad92936,plastic,PC
-3ceeb80e-a668-468f-9faf-e319be44c649,metal,Aluminium
-44a8f21b-77c8-4567-8b73-a3b5611747e6,metal,Copper
-95ba68b8-dcfa-4bcd-891c-0f40056696a4,metal,Copper
-5504775d-983e-46ed-85bb-9ef8c247f426,metal,Steel
-5e2c71fe-d3a8-426f-9e8b-75e4a2cc1832,wood,Teak
-2ed93ef0-434c-4d30-92da-b04e14d2a004,wood,Oak
-646d8f05-f165-454f-ba01-91ab34e62ad9,wood,Pine
-4c919a20-358e-4d6e-845a-4c0e5f8bf6ff,plastic,PE
-0db98652-5b9a-43c0-929b-81815d3f6ec3,wood,Oak
-f549dff5-a0a6-4702-85ed-46abe3486490,metal,Steel
-33260255-b1bc-449a-a404-a5f5f039e387,plastic,PC
-9f2dad6f-5f0b-4128-ba65-79f10d5abf5a,plastic,PE
+a93d06e7-bb88-4fd2-a401-e02f1330b873,plastic,PVC
+3469cc7b-d108-40b7-b800-d9739aa0e4f9,plastic,PE
+2ccb1c94-2293-4cb7-b0c7-9cae08203b1f,plastic,PC
+4000b7dc-cf99-4bc5-80e8-ebe3353342ce,wood,Oak
+7d3109aa-6129-4432-8e61-6ba0c740d23b,wood,Pine
+5bb14b26-1c8b-4fd0-bdf1-39a1d6a744f0,wood,Teak
+a81640ef-f8f2-44d3-ba59-913aef97eadb,metal,Aluminium
+ea5c757b-836b-4b68-b251-d087f4929c88,metal,Steel
+f1fb6bd9-700f-4a43-af9d-e88dc02e9c62,metal,Copper
