--- output/Material_t2.csv
@@ -1,101 +1,0 @@
-id,material_type,material
-c21a72c4-8f91-492d-a423-ffcd39a1251d,plastic,PC
-4380c4ca-c38a-4008-b157-a75f4488990c,wood,Teak
-c9c4f35a-28e1-4be3-bdb8-1fe40d4c0a1f,metal,Aluminium
-d0272ddf-f5e9-46b4-a784-d911b8857139,wood,Oak
-6e9791a9-5723-4d60-be39-cc49c3deb833,metal,Steel
-aef2e3d8-57ee-497a-abd3-65f50c3d8867,metal,Steel
-68e3b31d-29f4-469d-8f8b-327d0b7f67f2,metal,Copper
-1f7200e4-f8b6-44d9-abcb-0ba9d298a57b,wood,Teak
-73b3dd93-fed2-430e-95d1-dc7ff0fdf0ca,wood,Pine
-c16961c1-3395-43d2-8534-6bd0cfacf0da,wood,Pine
-8a904f4f-215a-4b23-9f22-48005ec54a41,plastic,PVC
-73e5e2e4-3437-4840-b16b-c1b3a348ff2c,metal,Aluminium
-46455798-6231-4c6a-8b29-bf07ed71e86a,plastic,PE
-04e960d9-c6cd-44a2-94bd-326511f82749,plastic,PC
-f5716aec-1760-4868-ba5a-781728514eca,wood,Pine
-90f9d04c-680c-49ed-9060-2630bd923bee,plastic,PC
-071dca39-d277-4b3c-aa2a-b47ac960872e,wood,Oak
-40fbf449-2138-45d1-8e62-1c9c24911faf,wood,Pine
-3fc06c75-e2ec-4fe1-9904-c674c4ef90d0,metal,Aluminium
-61ba1bf5-94f0-46ff-b89f-a4d6974dffbd,wood,Oak
-51fb2bec-6c6f-420d-9516-fd5d1e705306,wood,Teak
-9b502fde-858b-4aa3-808d-1c78deaa36b3,plastic,PVC
-2ac64273-5b4a-4be2-8270-961caa09a8fd,metal,Steel
-262d58dd-c745-446b-9394-7774ebac9bc0,plastic,PE
-da74fd20-d1bb-42a2-98e9-eb9d23aff201,wood,Pine
-074725cb-7cc4-400c-aea4-9155bb1e3de3,plastic,PC
-a2a8d0ee-5ec3-435c-8edf-b69b0a2cbc5d,plastic,PVC
-0aad618b-4649-41be-bf97-ec9054bdf2a6,wood,Oak
-a67c38df-f578-432b-a645-ce8f1fefe674,wood,Teak
-01498495-a458-48f6-86cf-cb92787fcac8,wood,Oak
-02d55c7e-3698-477f-a5a1-ab4a9953db20,metal,Copper
-f85a10e3-daa4-4350-b17e-e59a10573756,wood,Teak
-fb69528f-8ce2-4cee-8835-6e6921f915b1,metal,Aluminium
-f0e37fce-893c-456b-b59b-ee990720d989,metal,Copper
-db2d93cc-053d-4c8b-bf2a-3bf916b23a95,wood,Teak
-fabbb50f-80be-4ad3-9411-e62cc52ab469,wood,Oak
-f3328fa0-dc8f-4720-9a62-e398ee3cf899,wood,Oak
-1d95ef45-f8e7-4a11-b5b5-b7a7df12e601,metal,Copper
-ece1c3b4-6eec-4704-83af-d6e0dd8993e2,plastic,PVC
-eba50d3f-eb02-4611-ba66-f6c5ed7db49c,wood,Oak
-8686017d-1f8e-4643-a2a3-5697e071c8cf,metal,Copper
-8442b37f-17c3-4eef-8925-b7cb6145cee5,plastic,PC
-46ed09d0-724d-40ac-88e7-b4259b4a3b8d,wood,Oak
-bdcb96f2-6a79-43c4-8f3e-56740f68160e,wood,Pine
-f49f663c-e593-462a-80e0-fae056bf774f,metal,Steel
-4078295f-573d-4522-9af4-1663194cc03b,plastic,PE
-038c8a28-4ca8-497c-a562-c4bfe791612b,wood,Teak
-fcd839d1-bb0c-4b64-897b-92ba67607ad0,plastic,PC
-3792d7e4-cf2b-47e7-96ea-f39eb9ba9b50,plastic,PVC
-b0b56d72-474c-4062-a1fb-31415ae36900,wood,Pine
-203caf0e-65a6-4957-8a67-6b66e04dca35,wood,Pine
-505df589-48ce-4fc9-94fb-f77fcbc1e0be,wood,Oak
-a25ed9bf-f133-4dda-8407-1e0aa63d63d0,wood,Teak
-f549e49d-8f4f-47f3-b418-e87b7d635b8e,metal,Steel
-d6d3c0fc-1c0c-4599-b848-25007920fdd1,metal,Copper
-2e889978-7065-46c9-9d6f-ec1becc658bf,plastic,PVC
-afaae29a-520f-428d-a17c-e8975776a317,wood,Oak
-329de0c8-bab9-40cc-996b-31d29cc0e517,wood,Teak
-c4f3f127-97fc-4467-a899-43f3b8332cf2,metal,Steel
-a005522e-b69e-43e0-b56c-6c6f95663826,plastic,PVC
-dcb9b2a8-e242-4868-b0db-61fad7b5e5c7,metal,Steel
-4c414ab2-ec9f-4568-a35e-df53fdaec574,plastic,PC
-9029ffdf-b2a6-4bd3-9265-49c113c9124a,plastic,PE
-df5df931-cb87-427f-a527-3098890b9874,plastic,PE
-4ed35491-d8e2-4f68-8683-f0f9638d0d79,plastic,PE
-455b2524-b27c-482c-be64-04bf84505af7,wood,Teak
-210060e0-851e-4493-be1d-6c70732eb2f9,metal,Aluminium
-b8aea0a8-b15b-43a5-b9ad-dfb20e35e6bd,plastic,PC
-7e965461-a812-4548-bd12-483999fb21df,metal,Copper
-314b0cb4-7843-49ed-86d4-38468f5dabb9,metal,Copper
-8f07fab2-d1af-48a5-826f-4b7f53f10331,metal,Aluminium
-1ae5a188-72cf-4513-8255-1d58b74a7314,plastic,PVC
-e306a46c-7614-42f8-845c-d97f6c9aa60d,wood,Oak
-7bf9d859-6a7a-4ce9-ba33-8adc0e3834f2,plastic,PC
-892a4a1c-a26d-4ad9-81d0-6855ca09ca9c,metal,Copper
-3e153311-a02e-4cc8-9adf-ed43d8eb5ee3,metal,Steel
-69914551-279c-480a-87db-b924afa28c30,plastic,PE
-52e64386-04be-408e-b84a-1e3da8149a91,wood,Pine
-7cf1574f-3a87-4a9c-b42b-94d87c6602aa,metal,Copper
-8280598f-3e81-457c-bec2-04bdc759abd5,metal,Aluminium
-f28b2d39-add3-4bbf-a8ec-1f328f22053c,plastic,PVC
-ad1810d6-b9c8-43fc-a80e-5bde7d0ba618,metal,Steel
-7f5e6222-f3fc-42c8-9d29-1b3867323f4c,wood,Pine
-d1d295b7-ef99-41a9-9751-048184a9c9a4,plastic,PVC
-62c996b0-c4c5-4bb7-8dfc-3931c2269077,plastic,PE
-5bb719ee-b4ab-4596-9eea-8160c364e735,wood,Teak
-48cfc613-867a-4b98-8fe4-374016aadf16,wood,Oak
-aa00e9aa-6689-42dc-a1b8-ac8cf31c9e21,metal,Copper
-bdfe92bb-db68-4bd5-892c-28f3ba448606,metal,Aluminium
-cd90e0ed-dc00-4ed5-aadb-b5791acb0ce4,plastic,PVC
-f0345528-4464-407f-be87-8c2e3a9dc1eb,wood,Teak
-0e5977c6-0229-4953-a957-7414968078c9,plastic,PE
-89bb6217-24fb-436d-8115-4067d0953d50,wood,Teak
-2a7da960-1f9c-4477-9e3d-55af66d94f30,metal,Steel
-a14d1406-15bc-4ceb-adfd-85cd0da57187,wood,Pine
-092a7444-5e11-4c76-ad33-ffec8a84ee83,plastic,PVC
-1631f475-c3ca-43ed-ad62-e36cb4d78b4c,plastic,PC
-0f0aa266-13e9-407d-989a-0a0fb22281cf,metal,Aluminium
-123fe829-b7dd-42ce-8efa-ab341fb38d42,plastic,PC
-1f58466b-f221-47d3-8148-3152b0613122,metal,Aluminium
